requirements.txt
# Flask and extensions
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# CUPS and printer discovery
pycups==2.0.1
zeroconf==0.131.0

# PDF generation
reportlab==4.0.8

# HTTP client
requests==2.31.0

# Utility packages
python-dateutil==2.8.2
typing-extensions==4.9.0